test: move FoxESS/FoxCloud support_discharge_freeze check into test_inverter.py
Wasn't Teslemetry-specific - springfall2008 flagged it on PR review.
